requirements.txt
streamlit>=1.28.0
boto3>=1.34.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-pptx>=0.6.21
